requirements.txt
pandas
pydantic
openai
fastMCP
openai